refactor: extract JiraProjectOption.swift from JiraExportConfig.swift
Splits the `JiraProjectOption` lookup value out of JiraExportConfig.swift.

Byte-identical move of lines 70-86. Both files import Foundation.

JiraExportConfig.swift: 92 → 74 lines. Tests: 667.
