+++ Sources/BugNarrator/Models/JiraProjectOption.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+struct JiraProjectOption: Equatable, Identifiable {
+    let projectID: String
+    let key: String
+    let name: String
+
+    init(projectID: String? = nil, key: String, name: String) {
+        self.projectID = projectID ?? key
+        self.key = key
+        self.name = name
+    }
+
+    var id: String { projectID }
+
+    var displayLabel: String {
+        "\(key) - \(name)"
+    }
+}
